requirements.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
django-cors-headers==3.4.0
django==3.0.8
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyjwt==1.7.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
